Add files via upload
an interactive Python project that visualizes two divide-and-conquer algorithms: Closest Pair of Points and Karatsuba Integer Multiplication. Users can select algorithms and input files via a simple UI. The project visualizes the closest points on a graph and displays the Karatsuba multiplication process as a tree. It auto-generates input data and uses Matplotlib and Graphviz for engaging visualizations, enhancing learning and debugging.